+++ src/modules/notes/Note/components/Dialogs/DialogDeleteOneNote/DialogDeleteOneNote.tsx
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useNoteContext } from "modules/notes/Note/Context";
 import DialogDeleteOneNoteView from "modules/notes/Note/components/Dialogs/DialogDeleteOneNote/DialogDeleteOneNoteView";
 import { deleteOneNoteService } from "modules/notes/Note/components/Dialogs/DialogDeleteOneNote/services";
-import { useDialogNote } from "modules/notes/Note/shared/hook/useDialogNote";
+import { useDialogNote } from "modules/notes/Note/shared/hooks/useDialogNote";
 import React, { memo } from "react";
 import { NOTE_QUERY_KEY, MENU_QUERY_KEY } from "shared/constants";
