--- src/modules/notes/Note/shared/hook/useDialogNote/types/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./useDialogNote.types";
-export * from "./useDialogNote.types";
