+++ src/modules/notes/Note/shared/hooks/useDialogNote/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-dialog-note.types";
